# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoxForge API",
    "description": "Text-to-speech synthesis API with voice profile management",
    "version": "1.0.0"
  },
  "paths": {
    "/api/synthesize": {
      "post": {
        "tags": [
          "synthesis"
        ],
        "summary": "Synthesize text to audio",
        "description": "Convert text to audio and return the generated file.\n\nRoutes to XTTS v2 (voice cloning) if the profile has a voice sample,\notherwise uses Edge-TTS (Microsoft neural voices).\nAutomatically cancels if the client disconnects mid-generation.",
        "operationId": "synthesize_text_api_synthesize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SynthesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/character-synth/extract-characters": {
      "post": {
        "tags": [
          "character-synth"
        ],
        "summary": "List character names found in text",
        "operationId": "extract_chars_api_character_synth_extract_characters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractCharactersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractCharactersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/character-synth/synthesize": {
      "post": {
        "tags": [
          "character-synth"
        ],
        "summary": "Synthesize text with character-cast voices",
        "operationId": "cast_synthesize_api_character_synth_synthesize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CastSynthesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CastSynthesisRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "cast": {
            "items": {
              "$ref": "#/components/schemas/CharacterMapping"
            },
            "type": "array",
            "title": "Cast"
          },
          "output_format": {
            "type": "string",
            "title": "Output Format",
            "default": "mp3"
          },
          "speed": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 50.0,
            "title": "Speed",
            "default": 100
          },
          "pitch": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": -10.0,
            "title": "Pitch",
            "default": 0
          },
          "volume": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Volume",
            "default": 80
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          }
        },
        "type": "object",
        "required": [
          "text",
          "cast"
        ],
        "title": "CastSynthesisRequest"
      },
...
      "CharacterMapping": {
        "properties": {
          "character": {
            "type": "string",
            "title": "Character"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "voice_id": {
            "type": "string",
            "title": "Voice Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "character"
        ],
        "title": "CharacterMapping"
      },
...
      "ExtractCharactersRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ExtractCharactersRequest"
      },
      "ExtractCharactersResponse": {
        "properties": {
          "characters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Characters"
          }
        },
        "type": "object",
        "required": [
          "characters"
        ],
        "title": "ExtractCharactersResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SynthesisRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500000,
            "minLength": 1,
            "title": "Text"
          },
          "voice_id": {
            "type": "string",
            "title": "Voice Id",
            "description": "Voice ID (e.g. es-ES-AlvaroNeural)"
          },
          "output_format": {
            "type": "string",
            "title": "Output Format",
            "description": "mp3 | wav | ogg | flac",
            "default": "mp3"
          },
          "speed": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 50.0,
            "title": "Speed",
            "description": "Speed in % (50-200)",
            "default": 100
          },
          "pitch": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": -10.0,
            "title": "Pitch",
            "description": "Pitch in semitones",
            "default": 0
          },
          "volume": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Volume",
            "description": "Volume in %",
            "default": 80
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id",
            "description": "Profile ID (optional)"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "album": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "track_number": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9999.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Number"
          }
        },
        "type": "object",
        "required": [
          "text",
          "voice_id"
        ],
        "title": "SynthesisRequest",
        "description": "Text-to-speech synthesis request."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}